fix: minor issues
