+++ .gitignore
@@ -2,6 +2,7 @@
 node_modules/
 .env
 .env.*
+!.env.example
 dist/
 coverage/
 .pnpm-store/
